Remove ringback in runtime
