+++ frame/dvm/evm/precompiles/transfer/src/tests/mock.rs
@@ -365,5 +365,4 @@ pub fn new_test_ext(accounts_len: usize) -> (Vec<AccountInfo>, sp_io::TestExtern
 	(pairs, ext.into())
 }
 
-pub type RingAccount = <Test as darwinia_evm::Config>::RingAccountBasic;
 pub type KtonAccount = <Test as darwinia_evm::Config>::KtonAccountBasic;
